(BSR)[PRO] fix: move ThumbCell component to collective offer table folder
